+++ data/resources/CHANGELOG.md
@@ -12,6 +12,7 @@ August 5, 2021
 - dev tools: Added friendly tooltip to zone names in WindowInspector
 - Evaluating Spawn.ID on a NULL Spawn will now return 0 instead of a parse error. This is intended as
   a convenience for utilizing Spawn variables that are set via ID.
+- Fix auto scroll menu item in console window (#140).
 
 July 31, 2021
 - Added /mqsettings command to toggle the MacroQuest Settings window (you can still get to it from the EQ button as well)